Dismiss the meeting invitation message view when user responds
The message detail view for a meeting invitation is now dismissed,
returning to the Hot view or the message list view, when the user taps
one of the Attend/Maybe/Decline buttons.  A meeting cancelation notice
is also dismissed when tapping the "Remove from calendar" button.

Links in the location field, including addresses, are now clickable in
the message detail view (when part of a meeting
request/response/cancelation).  Links were already clickable in the
event detail view.  This change extends the behavior to the message
detail view.

Fix a minor layout issue for the location field in the event detail
view.  The location was a few pixels too far to the right.

nachocove/qa#857