Better job of synching existing device contact
When an existing device contact is changed, do a better job of
updating the corresponding app contact to match the device contact.
In the past, existing email addresses and phone numbers were never
changed or deleted from the app contact; new items were simply added
to the collections.

Improve the display of birthdays and anniversaries for contacts.
Don't show a year, just the month and day, when the device contact did
not specify a year for the birthday.

Fix nachocove/qa#566