Don't crash when a McAttendee has an email with bad syntax
HockeyApp contains a crash where class Uri threw an exception parsing
a mailto: URI where the email address came from a McAttendee.  I was
unable to reproduce the problem, because the app does rather thorough
checking that email addresses are syntactically valid.  So I added an
exception handler based on the stack trace in the crash dump.  The
handler should avoid the crash by ignoring the attendee.

Fix nachocove/qa#973