Don't let the calendar month get jammed
The UIPanGestureRecognizers for the calendar view were ignoring cancel
events for the gestures.  This could leave the view in a bad state,
with the gesture recognizers all disabled, resulting in the month view
not responding to any UI gestures.  Fix the gesture recognizers to
look for the canceled events, treating them the same as normal end
events.

Fix nachocove/qa#727