Store attendees as a serialized string instead of McAttendee objects
Storing meeting attendees in the database as individual McAttendee
objects was taking time and space with no real benefit.  Change the
app to store attendees as a serialized string in the McCalendar (or
McException or McMeetingRequest) object.

McAttendee objects are still created and are still used by the UI
code when viewing or editing a meeting.  But they are now created from
the serialized string and are stored back into the serialized string.
McAttendee objects in memory no longer have a corresponding object in
the database.

Existing meetings are not being migrated when the app is upgraded.  It
was feared that that would take too long.  Instead, attendees will be
migrated to the new format when each individual meeting is saved.  The
McAttendee table will remain in the database; it will be empty in
fresh installs and will slowly shrink towards being empty in upgraded
apps.