On remote notification, wait for all accounts to sync
When the app was woken by a remote notification, it would start a
quick sync of all the accounts on the device.  But it wait for the
sync to complete only for those accounts listed in the notification.
When those accounts were done, the app would shut down, even if other
accounts were still in the process of synching.  This could result in
ActiveSync accounts losing messages, as described in nachocove/qa#1043

Change the code to wait for all accounts to complete their quick sync
before shutting down the app.  This doesn't resolve issue 1043, but it
should reduce the frequency with which it happens.