Fix display of device events that repeat yearly on day of week
Some events that came from the device calendar that repeat yearly on a
particular day of the week would incorrectly show the month of their
occurrence as January.  (This only affected the "repeats on" message
in the event detail. The actual date of the event was correct.)  That
problem is now fixed.