Create and edit simple events on the device calendar
Add support for creating and editing simple events on the device
calendar.  Simple events are ones that are not recurring, have no
attendees, and have no attachments.

A new event is added to the device calendar instead of the current
account when the current account does not have the capability to write
to its own calendar.