Ignore calendar exceptions that are too old to show on the calendar
The app lets the user scroll the calendar only one month into the
past.  Exceptions to recurring calendar items that are more than a
month in the past will never be shown on the calendar.  So there is no
good reason to store them in the database.  The app now ignores them.