Improve the quality of the results when searching contacts
Rewrite the code at the core of contact search.  Separate the code
that searches for an e-mail address from the generic contact search.
For both scenarios, weed out duplicate entries and sort the results by
how well they match the search string.

nachocove/qa#1027 This update mostly fixes the issue. But there is
more to do, mostly cleaning up the code, so I am not resolving the
issue just yet.