Limit the number of contacts considered when searching contacts
When searching for a contact or for an e-mail address, configure each
subquery to return no more than 100 items.  This applies to both
database queries and index queries.  Because each search word is
looked up in multiple places, it is possible for the overall search to
still return more than 100 items.

This change improves performance when searching for a short string
that matches lots of contacts.  It comes at the cost of not always
finding the contact that the user is looking for.  But if there are
more than 100 contacts in the search results, the chance of the right
contact being near the top of the results is small anyway.