UIDatePicker views need to have a correct size
Before iOS 9, UIDatePicker views always had a fixed size.  In iOS 9
the views are now resizable.  That means the old code that gave the
view a height of 0 is now broken.  Fix EditEventViewController to give
UIDatePicker views the correct height.

This deals with one of the problems listed in nachocove/qa#1214.
Other problems still need to be dealt with before the issue can be
resolved.