Fix AutoresizingMask for several subviews to fix layout issues
Starting with iOS 9, the main View for a ViewController with a
navigation bar changes size and location in between ViewWillAppear()
and ViewDidAppear().  The AutoresizingMask for some subviews needs to
be set correctly so that the subviews will react correctly to this
size change.  (In the past the AutoresizingMask didn't matter, because
the main View was not resized after the views were layed out.  Now
that the main View can be resized, we need to pay more attention to
the AutoresizingMasks.)

This fixes layout issues with the toolbar in the message detail view,
and with the initial position of the event detail view and the event
editor view.

Fix nachocove/qa#1214