Include the server type in the device calendar names
Users who configure their Calendar app to sync calendars from several
different servers tend to have calendars with similar names, such as
"Birthdays", "Holidays", or "Work".  Nacho Mail was showing only the
simple names, making it hard for the user to figure out which calendar
was linked up with which server.

Change Nacho Mail to include the server type (technically, the
EKSource.Title field) in the name of the calendar that is shown to the
user.

Fix nachocove/qa#1223