Allow the user to choose the account when creating a new event
When the user is creating a new event, allow the user to choose any
calendar in any account as the home for that event, rather than
limiting the event to a calendar within the current account.

When editing an existing event, the user cannot change the account,
only the calendar within the account.  This behavior is unchanged.

Fix nachocove/qa#1222