Stop discarding contact search results
When the user is searching contacts (either the general contact search
or autocomplete for a email address), the app would throw away a set
of search results and not display them if the user had already typed
another character before the results were ready.  If the user could
type quickly, then he might not see any results until he was done
typing.

Fix the contact search code to not throw away search results.  The
results are now displayed even if the user has typed in more
characters.  This is a better user experience, because it lets the
user know that the app is searching, and it gives the user some useful
information to look at while waiting for the final results.

This change improves the situation for nachocove/qa#1232.  It may or
may not be a complete fix for the issue.