Some improvements to the Android event detail view
Improvements to the event detail view on Android: The title looks more
like a title.  The description is displayed (if it is not in RTF,
which it often is).  The notes field is correct (except that there can
never be any notes, since there is no way to enter them).  The
organizer is displayed correctly.

Change some more local notification methods from assertion failures to
not-yet-implemented log messages.

Fix the launching process so that the user doesn't have to enter a new
account every time the app starts.