More Android event detail improvements
Show the attendees summary: the initials and first name of the first
four or five attendees.  There is still no way to go to the attendees
view to see all the details about the attendees.

Show the number of attachments, but not the information about the
individual attachments.

The event detail view now has all the fields that it is supposed to
have.  But some of the fields have incomplete information (description
and attachments), and there is no way to change information or to
segue to a screen with details for a field (reminder, attendees,
notes).