Correctly localize dates and times
Localization of dates and times within the app was inconsistent.  When
using the app in a locale with 24-hour time, the app would sometimes
display the time correctly as "16:28", but other times would show it
as "4:28 pm".  There were similar problems displaying dates.

Clean up the code that displays dates and times to always localize
them correctly.  The methods in class Pretty for displaying dates are
times were rewritten to have more consistent names and more correct
behavior.  All uses of DateTime.ToString(string format) in the app
were changed to use the methods in class Pretty if the result was
displayed to the user.

(A couple of other compiler warnings unrelated to date/time were
fixed.)

Fix nachocove/qa#415