Cleanup event related activities and fragments
Make the event detail view its own Activity rather than just a
fragment that is added to another activity's view.  EventViewFragment
still exists, in case we want to reuse that UI somewhere else.  But it
is not wrapped in EventViewActivity.

Change CalendarActivity so that EventListFragment is hard coded in its
layout, as opposed to adding EventListFragment to the view hierarchy
dynamically.