Fix the account switch action when on the calendar tab
An earlier update of mine to clean up some of the activity/fragment
code broke the account switcher when the app was on the calendar tab,
causing the app to crash when the account switch button was tapped.
This change undoes some (but not all) of my previous update so that
the account switcher is again working correctly on the calendar tab.