Attendees view, round 1
Create the beginnings of the attendees view.  The view doesn't do
anything, and it just shows the attendee email addresses in an ugly
list.  The attendee view will be completed in a later update.

This update hooks the attendee view into both the event detail view
and the event editor view.  That integration is complete (except for
the editor processing the changed list of attendees).