Fix many of the crashes when rotating the device
Create class NcActivityWithData<T> to be used by activities that need
to preserve data across the OnDestroy()/OnCreate() calls that happen
upon configuration changes such as rotating the device.  Change
existing activities to use this for data that is stored in
IntentHelper.  For some activities that can now support rotation,
remove the code that disabled rotation.  For other activities that
still crash on rotation for other reasons, add the code that disables
rotation.

With these changes, most views in the app will not crash when the
device is rotated.  I tested most of the views other than the login
and new account views.  The app will still crash when launched in
landscape mode.  The behavior is not always correct when the device is
rotated.

Improves the situation for nachocove/qa#1360, but is not a complete
solution.