More fixes for fragment lifecycle when rotating the device
More fixes for handling fragments, avoiding the creating of duplicate
fragments when rotating the device.  This update fixes most of the
activities that have a single non-dialog fragment.  Message list views
can now be rotated.  If the contacts view has a search in progress,
that search will be preserved when the device is rotated.

nachocove/qa#1360