Fix quirky behavior when creating an all-day event late in the evening
When the user creates an all-day event when the current local time is
between 10:30pm and 11:30pm, and the user doesn't make any other
adjustments to the start or end times, then the event will be two days
long.  It is more likely that the user wanted it to be only one day
long.  Fix the code (on both iOS and Android) to recognize this
unusual situation and make the event be only one day long.

Fix nachocove/qa#1370