Prevent the sending of empty support requests
On both iOS and Android, the contact and description fields must be
non-empty before the support request can be sent.  (They don't have to
have meaningful information; they just have to have something.)

On Android, fill in the contact field with the e-mail address of the
current account if there is one.  (This was already the case on iOS.)

On Android, don't crash if the user presses the back button in between
hitting send and when the "Message Sent" dialog pops up.

Fix nachocove/qa#583
Fix nachocove/qa#1380