Make event notifications high priority
Event notification are given the correct category and are marked as
high priority.  (High priority is supposed to trigger a heads-up
notification, but I can't get that to happen.  Not sure why, and
Android docs aren't very helpful.  But this is obviously the right way
to set the priority, so I am checking in the change.)