Support the Android device calendar
This is done very differently from iOS.  Synching the device calendar
with the app is problematic because Android does not keep a "last
modified" field for events, so there is no feasible way to figure out
which events need to be synched.

Instead, events are read from the Android calendar database as needed
when showing the calendar list view.  When the user want to view an
event detail or create a new event, an activity is started in the
Android calendar app.

The code is mostly complete.  But there are still some edge cases,
limitations, and bugs that need to be dealt with.  Fixing those issues
will happen after it has been decided whether or not this is the
correct approach for handling the Android calendar.