View Android device events in the app
Events from the device calendar are now viewed in Nacho Mail rather
than in the Android Calendar app.  When opening the event detail view
for a device event, an in-memory McCalendar object is created and
populated with information from the device event.

Currently the user can't make any changes to the event.  That will
come later.