Make class DataFragment be non-generic to avoid a Xamarin limitation
Xamarin does not allow generic C# classes to be instantiated from Java
code.  (To me, that's a perfectly reasonable limitation.)  That means
activity and fragment classes can't be generic.  Therefore, change
class DataFragment from a generic class to a regular class with
generic methods.

Fix nachocove/qa#1434
(I hope it is fixed, but we can't reproduce the problem so I can't
verify the fix.)