Create back stack when app is launched by an event notification
When the app is launched by tapping an event notification, insert a
calendar view activity into the activity back stack behind the event
detail view activity.  This keeps the user within the app when tapping
the back button rather than returning the user to the device's home
screen.

(Figuring out whether or not the app is being launched is not trivial.
I am not sure about the accuracy of the code that I have.)