Don't leak brain index locks
A code path involving reindexing a contact could result in the brain
task shutting down while an index lock was still held.  This would
result in a crash when the brain was restarted because the lock was
released from a different thread than the one that acquired the lock.
This has been fixed by being more diligent about releasing locks.
(See nachocove/qa#1467 for more details.)

Don't release all the index locks (and close all the index writers)
when only one account needs to be cleaned up.

Fix nachocove/qa#1467