Allow changing the device calendar that owns a device event
When creating a new event, allow the user to select one of the device
calendars as the owner of the event.

When editing an existing event on one of the device calendars, allow
the user to move the event to a different device calendar.  (But like
iOS, the event cannot be moved to a different account, only a
different calendar within the same account.)

Allow the user to change the reminder time for a device event when
editing the event.  (Changing the reminder for device events is still
not implemented yet from the event detail view.)