Attend/maybe/decline buttons on event detail for device events
On both iOS and Android, when the user taps one of the
attend/maybe/decline buttons on the event detail view for an event
that is owned by the device calendar, pop up a dialog explaining that
these buttons don't work for events from the device calendar.  (This
used to be the behavior on iOS, but a regression caused the buttons to
silently do nothing.  On Android, this was unfinished code.)

On Android, allow the user to change the reminder for an event from
the device calendar from the event detail view, without having to edit
the event.