Several Android calendar fixes
Don't crash when a meeting cancelation message is opened in the
message detail view.

Implement the "Remove from calendar" button in the event detail view.
(It was accidentally forgotten when the view was implemented.)

Don't crash when deleting a meeting that was created in the AmazonMail
web UI.  The Amazon Exchange server, unlike all of the other Exchange
servers the app interacts with, leaves the OrganizerEmail field unset
for meetings where the current user is the organizer.  The code was
expecting that field to always be set.  Change the ActiveSync code to
always set the OrganizerEmail field.
Fix nachocove/qa#1497