Start the back end during PerformFetch and RemoteNotification
A change not too long ago removed the BackEnd.Instance.Start() call
from the normal startup process, moving the call to when the app comes
into the foreground.  That meant it was never called during QuickSync
mode, so the app wasn't never synching anything unless it was in the
foreground.

Explicitly call BackEnd.Instance.Start() when starting a QuickSync.