Other Android apps can share with Nacho Mail
Add an activity with intent filters so that other apps can share data
with Nacho Mail, which will open the message compose view so the user
can send the shared data in an e-mail message.  If the shared data is
text, it will be part of the body of the new message.  If the shared
data is files, such as pictures, those files will be attachments on
the new message.