Allow iOS account settings page to scroll all the way to the bottom
If the settings page for an individual account was too tall for the
device's screen (which can happen on iPhone 4s, 5, and 5s), it was not
possible to scroll the view all the way to the bottom.  The delete
button was always off the bottom of the screen, making it impossible
to delete the account.  Fix the scroll view settings to avoid this
problem.