Avoid long scroll times in calendar when jumping to a particular date
Android list views have a maximum speed at which they can scroll.
Scrolling a long ways will always take a long time.  To avoid long
scroll times in the calendar view when jumping to a particular date,
change the code to instantly jump most of the way there and then
scroll the remaining portion.  The user still has the experience of
scrolling, but never has to wait for scrolling to complete.

This does not work if the month is pulled down and the month needs to
be changed as part of scrolling.  For reasons that I can't figure out,
the listView.SetSelection() call has no effect, so the view ends up
scrolling the entire way.  Because the month view is not often used,
and because the user experiences a delay, not a loss in functionality,
I am not taking the time to fix this right now.

Fix nachocove/qa#1576