Fast calendar scrolling, event when month view is open
Fix the problem where the "Today" button was taking a long time to
scroll the calendar list to today when the month view was pulled down.
Do this by delaying the calendar list scroll until after the month
view has been adjusted, so that adjusting the month view doesn't
interfere with the calendar list view SetSelection() call.