Make user choose an account when sharing with Nacho Mail
When the user shares something with Nacho Mail from a different
Android app and the user has multiple accounts in Nacho Mail, present
a dialog so the user can choose the account from which the message
will be sent.  If the user taps the back button, taps outside the
dialog, or rotates the device, the entire share operation is canceled.
(This dialog is shown only when there are multiple accounts.  If there
is only one account, the app goes straight to the message compose
view.  This dialog is temporary, and will be removed once the message
compose view allows the user to choose the account.)

When the user shares something with Nacho Mail from a different
Android app and there are no configured accounts, present a dialog
explaining that there are no accounts.  When the user dismisses the
dialog, the share operation is canceled.  (This situation will arise
if the user has installed Nacho Mail but never run it.)