Fix some annoyances with the Android credentials input fields
Tweak the text input fields on the Android credentials page to behave
as most user would expect:

When the credentials page become visible, focus is placed in the email
field and the soft keyboard is made visible.  The user no longer has
to tap in the email field to start input.

When pressing the next button on the keyboard to transfer from the
email field to the password field, the soft keyboard remains visible
rather than becoming hidden.  The user no longer has to tap in the
password field to make the keyboard visible again.

When pressing the done button on the keyboard after entering the
password, the keyboard disappears making the Submit button visible.
The password field is no longer a multi-line text field, and the user
no longer has to explicitly dismiss the keyboard in order to tap the
Submit button.