Don't crash when the user hits Enter in an empty attendee search field
When editing the attendees for a meeting, if the user hit the Add
button then hit Enter in the search field without typing anything, the
app would crash.  This has been fixed.  The app now behaves as if the
user canceled the add.

(The same bug happened on both iOS and Android.  The causes, and
fixes, were different.  But the symptoms were identical, so the fixes
are lumped together.)