ensure services/backend are started when needed
Dynamically starting the backend worked on app-start, but not on BG->FG
if the services were stopped in the BG (i.e. iOS). In that case, we
must also start the services.
We do NOT however, want to constantly call start on the backend, since
that just adds more CPU work that is not necessary.