set a timeout for an IDLE command
if the connection to the server hangs or otherwise becomes inoperative,
we could wait forever before we notice. So set a timeout on the IDLE
command (9 minutes for gmail, and 30 minutes for everyone else).
Also print folder names that are ‘standard gmail’ folders, as we’re not
printing any PII.