don't report comm results if we've been cancelled
In the case of a cancellation, further up in the code, we just resolve
any pendings as deferred, and return from the task (exiting it). But
because of the ‘finally’, we also tried to do ReportCommResult, looking
up the server (BEContect.Server is a lookup). Since this is part of an
account deletion, the server is now gone.
resolves nachocove/qa#1670