Don't crash when scrolling the month view to the future
In a past update I changed the behavior of
INcEventProvider.IndexOfDate().  I updated all of the Android uses of
that method, but forgot one of uses in iOS code.  This resulted in the
app crashing when the month pull-down on the calendar view was
scrolled more than five months into the future.

Fix nachocove/qa#1669