Avoid crash caused by using a disposed BodyView object
It is possible for the user to dismiss the message detail view while
the message bundle is being created in the background.  This would
result in a disposed BodyView object being used once the message was
ready to be rendered, which would cause a crash.  Add checks so that
the app detects this situation and doesn't try to use the disposed
BodyView.

Fix nachocove/qa#1672