Don't crash on iPad when presenting an action sheet
On an iPad, an action sheet dialog is anchored to a location on the
screen rather than being at the bottom of the screen.  The app now
specifies the anchor point for the action sheet, which keeps iOS from
killing the app.  (Because iPad is not a priority, the action sheet
dialog is not always in the best possible place.  But it should always
be visible.)

Change all the uses of UIAlertController in MessageComposeView to use
NcAlertView or NcActionSheet instead.

Remove the code in NcAlertView and NcActionSheet that was specific to
iOS 7, since that version of iOS is no longer supported.

Fix nachocove/qa#1684