protect against some null references
Not really sure what in this bug was null. My guess is that the task
was somehow deleted before (or during?) this call. The only way this
could cause a crash, is if something went wrong inside the catch
(Exception ex), which looks like it could only be task. So add some
protections against it, and other things while I’m at it.
resolves nachocove/qa#1714