Index text parts of all messages regardless of content encoding
Remove the content encoding check when deciding whether or not to
index the body of an e-mail message.  Attempt to index the text parts
of all messages, counting on MimeKit to handle any unusual encodings.

When indexing HTML, convert HTML entities (e.g. &nbsp; or &lt;) into
their respective characters before passing the string to the indexer.

Minor improvements to the performance of indexing e-mail messages,
such as using StringBuilder instead of string concatenation when
building up a string from lots of parts.

Fix nachocove/qa#1636