Index email messages right after they are synched
Tweak the scheduling of brain tasks so that email messages are indexed
shortly after they are synched.  (Unless they are in the junk folder
or the trash, in which case they will be indexed, or not, a little bit
later.)  This initial indexing will be headers only; the message will
be reindexed if its body is downloaded later.