Cache device calendar events on Android; load only when necessary
Cache the events that come from the Android device calendar.  Reload
them only when an AndroidOS calendar-changed event happens, rather
than every time the event list view is refreshed.  This should speed
up the initial display of the event list view.

One side effect is that the device events will not show up on the
calendar until several seconds after the app is launched.  During this
small window, the event list view will show app events but not device
events.

This change improves nachocove/qa#1750, but it is not a complete
fix.  I am leaving the issue open for now.