A couple of indexing fixes
When the processing of the brain's persistent queue is interrupted,
which happens when the back end is supposed to be abated, don't add
another persistent queue event to the brain's queue.  This will result
in a continuous loop of attempting to process persistent queue events
but never making any progress because of the abatement.  That will use
up all the CPU exactly when the brain is supposed to not use any CPU.
There isn't a good way for the brain to block until the abatement is
over, so any unprocessed persistent queue events will just wait until
the next periodic glean event.

Don't reindex e-mail messages when the IMAP server has merely changed
some flags on the message.