Fix "Too many open files" crashes on iOS due to leaking DB connections
NcTimer does not have the code that NcTask does for cleaning up
database connections when the task completes.  Any DB connections used
in a timer callback would only be cleaned up in the DB garbage
collector, which is scheduled to run every 15 seconds.  An iOS quick
sync often takes less than 15 seconds, which means the garbage
collector won't run.  If the user leaves the app in the background for
a long time, there could be a large number of quick syncs that create
a few new DB connections that are never cleaned up.  This would
eventually lead to the app's process running out of file descriptors.

Several changes were made to avoid this problem:

Change NcTimer to close the DB connection when the callback is done if
there are lots of open DB connections.  (This is what NcTask already
does.)  This change should stop DB connections from accumulating if
the DB garbage collector never runs.

Change NcTask to always close the DB connection when a long running
task completes.  The thread for a long running task is unlikely to be
reused in the near future, so there is no advantage to keeping the DB
connection open.  This change should help keep the number of open
connections lower.

When the app is in quick sync mode, set the initial delay for the DB
connection garbage collector timer to 1 second instead of 15 seconds.
(The repeating period is still 15 seconds.  Only the first run is sped
up.)  This change will run the garbage collector at least once per
quick sync, so stale connections should never accumulate over time.

There are also two other changes not directly related to the
too-many-open-files fix:

When the app is in quick sync mode, start the process monitor with an
initial delay of 1 second, so that a monitor report is produced during
each quick sync.

Get rid of the "NcTimer created" log message.  This message was always
followed by an "NcTimer set" message, so the "created" message
contained no useful information.