Keep the brain from permanently shutting down
If the NcTask cancellation token was cancelled before the brain could
process its TERMINATE event, then the brain would be disabled until
the app was killed/restarted.  Several changes were made to avoid this
problem:

When the brain starts, discard any TERMINATE or PAUSE events at the
head of the brain's queue.

In NcBrain.StopService(), add the TERMINATE event to the brain's queue
only if the brain is actually running.

Set NcBrain.IsRunning to false whenever the brain's task exits.

Fix nachocove/qa#1770