Basic message list filtering functionality
In all message list views that should support filtering (except for
the unified inbox) a spectacles icon in the button bar at the top
allows the user to cycle through the supported filtering options: all,
hot, focused, and unread.  The current filter setting is shown in a
text field at the top of the message list.

The UI is preliminary and needs improvement in a number of ways.
Support for filtering the unified inbox will be added in the near
future.