Improve the look of the bar that shows the current filter setting
When filtering is enabled for a folder, the current filter setting is
shown in a status bar at the top of the message list.  This change
improves the look of that status bar on both iOS and Android.  This is
not the final UI for this status bar, but it is a step in the right
direction.