Hopefully avoid DB connection GC race condition
When the app's process resumes after having been suspended for a
while, all existing database connections will appear stale until the
thread using that connection gets a chance to run again.  The existing
code could close an apparently stale connection just as its thread was
about to use it again, resulting in an SQLite exception and a crash.

To avoid this race condition, change the definition of "stale" in the
DB connection garbage collector.  To be stale, the connection needs to
have not been used since sometime before the previous time that the
garbage collector was run.  With this change, connections will not be
considered stale just because the process was shut down for a while.

Fix nachocove/qa#1812