Change NcTimer locks so Dispose() won't block for very long
NcTimer objects had one lock object, which was held while the timer
callback was running.  Because time callbacks can be long running,
this could result in NcTimer.Dispose() blocking for an extended period
of time waiting to acquire the lock.  Dispose() is sometimes called on
the main thread during shutdown, and therefore it must always return
quickly.

Change NcTimer to have two lock objects.  One of the locks is used
when creating, scheduling, and disposing the object.  The other lock
is used when invoking the callback function.  The callback lock will
guarantee that only one callback function is running at a time, but it
will not block the timer from being disposed.

This change should prevent the app's shutdown process from pausing or
hanging.  In some cases, the hang was long enough that the app was
getting killed before it could finish shutting down.