Reimplement abatement to be more efficient and effective
Change the way that abatement (which is when the UI or sync code asks
the back end to chill while important stuff is done) is implemented.
Rather than shutting down and later restarting, a background service
that honors abatement requests will now simply pause while the
abatement request is in effect.  The work to shut down and restart the
service was excessive for the short amount of time that the service
was supposed to chill, and the signals to stop/start were often
delayed until well after the abatement period was over.  Pausing the
background thread is more efficient and has a more immediate effect.

The way that NcBrain schedules periodic gleaning was simplified.

iOS view controllers now log only one telemetry UI event for each
lifecycle method (ViewWillAppear, etc.), instead of both begin and end
events.  The separate begin and end events were too close to each
other in time to be useful.  The end event was just clutter in the
telemetry.