Faster calendar load/refresh times, especially on Android
Keep a singleton cache of all the events that could be displayed in
the calendar list view.  Initialize the cache shortly after app
launch, and keep the cache up to date even when the calendar view is
not visible.  With this change, tapping on the Calendar tab should
result in the correct events being displayed almost instantaneously
(except when tapping the Calendar tab within the first few seconds of
launching the app).  This is a significant improvement for Android;
iOS already did this.
Fix nachocove/qa#1825

On Android, the calendar will automatically extend into the future as
the user scrolls the event list, the week view, or the month view.
The user is no longer limited to seeing just five months into the
future.
Fix nachocove/qa#1594