Keep contact interactions from slowing down the app
Fix an order of operations bug in the DB query in
McEmailMessage.QueryInteractions() that was making the query much
slower than it should have been.

Reorganize code so that McEmailMessage.QueryInteractions() is always
run on a background thread.  Even with the bug fix, it is expected
that the query will sometimes be slow enough that it should not happen
on the UI thread.

Fix nachocove/qa#1854